requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
astroid==2.15.6
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dill==0.3.7
exceptiongroup==1.1.2
execnet==2.0.2
Flask==2.3.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.2.0
pylint==2.17.5
PySocks==1.7.1
pytest==7.3.2
pytest-xdist==3.3.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
webdriver-manager==3.8.6
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0

